Removed inclusion of toolchain lib
There is probably a better way to do this, but I am not familiar enough with the embedded cross compile framework. My instinct says this should be something that can be enabled/configured in either project.cmake or board.cmake

When 

-L/Applications/ArmGNUToolchain/12.2.mpacbti-rel1/arm-none-eabi/arm-none-eabi/lib 

is added to the cmake, all programs we tested (including the simplest STM32Blinky) would not run on Nucleo (appeared to crash).

Tested across multiple projects and positioning this parameter in multiple locations - every program worked with this removed.